create useGenerateLeftItemsAlert hook
