+++ src/contexts/NewDelivery.tsx
@@ -1,8 +1,9 @@
-import { createContext, ReactNode, useEffect, useState } from 'react'
+import { createContext, ReactNode, useState } from 'react'
 import { SubmitHandler } from 'react-hook-form'
 import useNewDeliverySteps from '@/hooks/useNewDeliverySteps'
 import { NewDeliveryContextValues } from '@/interfaces/newDeliveryContextValues'
 import useSetStep3Items from '@/hooks/useSetStep3Items'
+import useGenerateLeftItemsAlert from '@/hooks/useGenerateLeftItemsAlert'
 
 interface NewDeliveryProviderProps {
   children: ReactNode
@@ -25,7 +26,7 @@ export const NewDeliveryContext = createContext<NewDeliveryContextValues>({
   setStep3Items: () => {}
 })
 
-interface MoveGood {
+export interface MoveGood {
   type: string // palettes | packages | pieces
   quantity: number
   zone: string
@@ -43,68 +44,13 @@ export default function NewDeliveryProvider({ children }: NewDeliveryProviderPro
   const [isExceedQuantity, setIsExceedQuantity] = useState(false)
 
   useSetStep3Items(formsData, step3Items, setStep3Items)
-
-  useEffect(() => {
-    const currentItems = { pallets: 0, packages: 0, pieces: 0 }
-
-    step4Items.map((item) => {
-      if (item.type === 'pallets') {
-        currentItems.pallets += item.quantity
-      } else if (item.type === 'packages') {
-        currentItems.packages += item.quantity
-      } else if (item.type === 'pieces') {
-        currentItems.pieces += item.quantity
-      }
-      return item
-    })
-
-    const leftItems = {
-      pallets: step3Items.pallets - currentItems.pallets,
-      packages: step3Items.packages - currentItems.packages,
-      pieces: step3Items.pieces - currentItems.pieces
-    }
-
-    if (leftItems.pallets < 0) {
-      setIsExceedQuantity(true)
-      setAlertMessage(
-        `Количеството на палетите е надвишено с ${currentItems.pallets - step3Items.pallets}`
-      )
-    } else if (leftItems.packages < 0) {
-      setIsExceedQuantity(true)
-      setAlertMessage(
-        `Количеството на пакетите е надвишено с ${currentItems.packages - step3Items.packages}`
-      )
-    } else if (leftItems.pieces < 0) {
-      setIsExceedQuantity(true)
-      setAlertMessage(
-        `Количеството на бройките е надвишено с ${currentItems.pieces - step3Items.pieces}`
-      )
-    } else if (leftItems.pallets === 0 && leftItems.packages === 0 && leftItems.pieces === 0) {
-      setIsCompletedMove(true)
-      setAlertMessage('Всички стоки са разпределени по зони')
-    } else {
-      setIsCompletedMove(false)
-      setIsExceedQuantity(false)
-
-      let palletsMessage = leftItems.pallets === 0 ? '' : `${leftItems.pallets} палети`
-      let packagesMessage =
-        leftItems.packages === 0
-          ? ''
-          : palletsMessage === ''
-            ? `${leftItems.packages} пакети`
-            : `, ${leftItems.packages} пакети`
-      let piecesMessage =
-        leftItems.pieces === 0
-          ? ''
-          : palletsMessage === '' && packagesMessage === ''
-            ? `${leftItems.pieces} бройки`
-            : `, ${leftItems.pieces} бройки`
-
-      setAlertMessage(
-        `Остава да поставите ${palletsMessage} ${packagesMessage} ${piecesMessage} в зони`
-      )
-    }
-  }, [step4Items])
+  useGenerateLeftItemsAlert(
+    step3Items,
+    step4Items,
+    setIsExceedQuantity,
+    setAlertMessage,
+    setIsCompletedMove
+  )
 
   function updateStep4Item(index: number, newItem: MoveGood) {
     const newStep4Items = [...step4Items]
