+++ src/interfaces/step3Items.ts
@@ -1,0 +1,5 @@
+export interface Step3Items {
+  pallets: number
+  packages: number
+  pieces: number
+}
